Added a framework independent top level testing module.
