+++ setup.py
@@ -1,4 +1,6 @@
 # Always prefer setuptools over distutils
+import io
+
 from setuptools import setup, find_packages
 # To use a consistent encoding
 from codecs import open
@@ -6,47 +8,51 @@ from os import path
 
 here = path.abspath(path.dirname(__file__))
 
-# Get the long description from the README file
-# try:
-    # import pypandoc
-    # long_description = pypandoc.convert('README.md', 'rst')
-# except(IOError, ImportError):
-#     long_description = open('README.md').read()
+BASE_DIR = path.dirname(__file__)
+README_PATH = path.join(BASE_DIR, "README.md")
+LONG_DESCRIPTION_TYPE = "text/markdown"
 
-long_description = open('README.rst').read()
+if path.isfile(README_PATH):
+    with io.open(README_PATH, encoding="utf-8") as f:
+        LONG_DESCRIPTION = f.read()
+else:
+    LONG_DESCRIPTION = ""
 
 flask_requires = [
-    'Flask>=1.0.2',
-    'graphene>=2.1.2',
-    'Flask-Graphql>=2.0.0',
+    "Flask>=1.0.2",
+    "graphene>=2.1.2",
+    "Flask-Graphql>=2.0.0",
+    "flask_unittest>=0.1.2",
 ]
 
 django_requires = [
-    'graphene-django>=2.0.0',
+    "django",
+    "graphene-django>=2.0.0",
 ]
 all_requires = flask_requires + django_requires
 
 tests_require = [
-    'coverage',
-    'pytest',
-    'pytest-cov',
-    'pytest-django'
+    "coverage",
+    "pytest",
+    "pytest-cov",
+    "pytest-django",
+    "parameterized",
 ]
 
 setup(
-    name='graphene_file_upload',
+    name="graphene_file_upload",
     packages=find_packages(exclude=["tests"]),
-    version='1.3.0',
-    description='Lib for adding file upload functionality to GraphQL mutations in Graphene Django and Flask-Graphql',
-    long_description=long_description,
-    long_description_content_type='text/x-rst',
-    author='Lucas McGartland',
-    author_email='lucasmcgartland@gmail.com',
-    url='https://github.com/lmcgartland/graphene-file-upload',  # use the URL to the github repo
-    # download_url = 'https://github.com/lmcgartland/graphene-file-upload/archive/0.1.0.tar.gz',
-    keywords=['graphql', 'graphene', 'apollo',  'upload'],  # arbitrary keywords
+    version="1.3.0",
+    description="Lib for adding file upload functionality to GraphQL mutations in Graphene Django and Flask-Graphql",
+    long_description=LONG_DESCRIPTION,
+    long_description_content_type=LONG_DESCRIPTION_TYPE,
+    author="Lucas McGartland",
+    author_email="lucasmcgartland@gmail.com",
+    url="https://github.com/lmcgartland/graphene-file-upload",  # use the URL to the github repo
+    # download_url = "https://github.com/lmcgartland/graphene-file-upload/archive/0.1.0.tar.gz",
+    keywords=["graphql", "graphene", "apollo",  "upload"],  # arbitrary keywords
     install_requires=[
-        'six>=1.11.0',
+        "six>=1.11.0",
     ],
     tests_require=tests_require,
     classifiers=[
@@ -75,9 +81,9 @@ setup(
         "Framework :: Flask"
     ],
     extras_require={
-        'flask': flask_requires,
-        'django': django_requires,
-        'all': all_requires,
-        'tests': tests_require,
+        "flask": flask_requires,
+        "django": django_requires,
+        "all": all_requires,
+        "tests": tests_require,
     },
 )
