+++ tests/test_flask.py
@@ -2,7 +2,11 @@ import json
 from tempfile import NamedTemporaryFile
 
 import pytest
+from flask.testing import FlaskClient
+from parameterized import parameterized
 
+from graphene_file_upload.flask.testing import GraphQLFileUploadTestCase
+from graphene_file_upload.testing import file_graphql_query
 from .flask_app import create_app
 from .schema import schema
 
@@ -27,7 +31,7 @@ def client():
     ),
     indirect=['client']
 )
-def test_single_file(client, file_text, expected_first_line):
+def test_upload(client, file_text, expected_first_line):
     query = '''
         mutation testMutation($file: Upload!) {
             myUpload(fileIn: $file) {
@@ -39,21 +43,14 @@ def test_single_file(client, file_text, expected_first_line):
     with NamedTemporaryFile() as t_file:
         t_file.write(file_text.encode('utf-8'))
         t_file.seek(0)
-        response = client.post(
-            '/graphql',
-            data={
-                'operations': json.dumps({
-                    'query': query,
-                    'variables': {
-                        'file': None,
-                    },
-                }),
-                't_file': t_file,
-                'map': json.dumps({
-                    't_file': ['variables.file'],
-                }),
-            }
+
+        response = file_graphql_query(
+            query,
+            op_name='testMutation',
+            files={'file': t_file},
+            client=client,
         )
+
     assert response.status_code == 200
     assert response_utf8_json(response) == {
         'data': {
@@ -63,3 +60,42 @@ def test_single_file(client, file_text, expected_first_line):
             },
         }
     }
+
+
+class MyUploadTestCase(GraphQLFileUploadTestCase):
+    app = create_app(schema=schema)
+
+    @parameterized.expand([
+        (u'Fake Data\nLine2\n', u'Fake Data'),
+        # Try the fire emoji
+        (u'\U0001F525\nLine2\nLine3\n', u'\U0001F525'),
+    ])
+    def test_upload(self, client, file_text, expected_first_line):
+        query = '''
+           mutation testMutation($file: Upload!) {
+               myUpload(fileIn: $file) {
+                   ok
+                   firstLine
+               }
+           }
+       '''
+
+        with NamedTemporaryFile() as t_file:
+            t_file.write(file_text.encode('utf-8'))
+            t_file.seek(0)
+
+            response = self.file_query(
+                query,
+                op_name='testMutation',
+                files={'file': t_file},
+            )
+
+            assert response.status_code == 200
+            assert response_utf8_json(response) == {
+                'data': {
+                    'myUpload': {
+                        'ok': True,
+                        'firstLine': expected_first_line,
+                    },
+                }
+            }
