+++ tox.ini
@@ -28,5 +28,5 @@ deps =
     py3{4,5,6,7,8,9}: pylint
 commands =
     -coverage erase
-    py3{4,5,6,7,8,9}: pylint graphene_file_upload
+    py3{4,5,6,7,8,9}: pylint graphene_file_upload --disable=C0209
     py.test --cov=graphene_file_upload -v tests/
